requirements.txt
asgiref==3.2.3
cat-sdk==3.1.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-allauth==0.41.0
django-appconf==1.0.3
django-cleanup==4.0.0
django-cropperjs==0.0.5
django-heroku==0.3.1
django-imagekit==4.0.2
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
oauthlib==3.1.0
pilkit==2.0
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.19
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
